hoist style guide
